pyproject.toml
[project]
name = "drf-bookstore-api"
version = "0.0.1"
description = "Bookstore API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "gunicorn>=23.0.0",
    "psycopg[binary,pool]>=3.2.10",
    "pydantic-extra-types>=2.10.5",
    "pydantic-settings>=2.10.1",
]

[dependency-groups]
dev = [
    "model-bakery>=1.20.5",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "ruff>=0.13.0",
]

[tool.ruff]
exclude = [
    "**/migrations/**"
]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D",
    "ANN",
    "ARG002",
    "EM101",
    "TRY003",
    "COM812",
]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.per-file-ignores]
"manage.py" = ["PLC0415", "TRY003", "EM101"]
"tests.py" = ["S101"]
"tests/**" = ["S101"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "bookstore.test_settings"
python_files = ["tests.py", "test_*.py"]
